allow UsaFacts to be 7 days stale, current limit is 5. (#187)
This adds 2 days to how stale we let USAFacts get. This is better than disabling the fetch because it'll resume working without us needing to change anything and gives us 2 days to fix it before breaking the fetch again. It'd still be a good idea to revert this PR when the issue is fixed.

See https://covidactnow.slack.com/archives/C011Z21ST8V/p1609898936002300